apps/indonesian_nlp/requirements.txt
# Core Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# Celery for async tasks
celery>=5.3.0
redis>=4.5.0

# NLP and Machine Learning
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.1
Sastrawi>=1.0.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Text processing and language detection
langdetect>=1.0.9
textblob>=0.17.1

# Data serialization and handling
joblib>=1.3.0

# Performance monitoring
psutil>=5.9.0

# Database and caching
psycopg2-binary>=2.9.0
django-redis>=5.3.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
faker>=19.0.0

# Additional utilities
requests>=2.31.0
tqdm>=4.66.0
click>=8.1.0